+++ Source/TestProject/Packet.h
@@ -20,7 +20,7 @@ public:
 	PROTOCOL GetHeader() const;
 	char* GetBody() const;
 	char* GetBuffer() const;
-	ULONG GetPacketSize() const;
+	size_t GetPacketSize() const;
 
 	bool IsValid() const;
 	void Clear();
@@ -31,14 +31,14 @@ private:
 	PROTOCOL header;
 	char* body;
 	char* buffer;
-	ULONG packetSize;
+	size_t packetSize;
 	bool isDAllocBuf;
 };
 
 class PacketInfo
 {
 public:
-	PacketInfo(char* const _buffer, ULONG _packetSize, PROTOCOL _header)
+	PacketInfo(char* const _buffer, size_t _packetSize, PROTOCOL _header)
 	{
 		buffer = _buffer;
 		packetSize = _packetSize;
@@ -46,12 +46,12 @@ public:
 	}
 
 	char* GetBuffer() const { return buffer; }
-	ULONG GetPacketSize() const { return packetSize; }
+	size_t GetPacketSize() const { return packetSize; }
 	PROTOCOL GetHeader() const { return header; }
 
 private:
 	char* buffer;
-	ULONG packetSize;
+	size_t packetSize;
 	PROTOCOL header;
 };
