+++ Source/TestProject/Session.h
@@ -11,16 +11,16 @@ public:
 
 	void ClearBuff();
 	void ClearBuff(bool isDeleteBuff);
-	void AddToBuff(char* const _buffer, const ULONG _packetSize);
+	void AddToBuff(char* const _buffer, const size_t _packetSize);
 
 	void SetUserId(const wchar_t* const _userId);
 	void SetBuffer(char* const _buffer);
-	void SetPacketSize(const ULONG _packetSize);
+	void SetPacketSize(const size_t _packetSize);
 
 	SOCKET GetServSock() const;	
 	wchar_t* GetUserId() const;
 	char* GetBuffer() const;
-	ULONG GetPacketSize() const;
+	size_t GetPacketSize() const;
 
 	bool IsValid() const;
 private:
@@ -28,5 +28,5 @@ private:
 	const char* ip;
 	wchar_t* userId;
 	char* buffer;
-	ULONG packetSize;
+	size_t packetSize;
 };
